requirements.txt
pandas
praw
nltk
psycopg2-binary
apache-airflow-providers-postgres